Refactor project structure: Move cursor file to utils package and migrate common config struct to common.go
