+++ internal/config/common.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package config
 
 import (
 	"context"
@@ -27,7 +27,7 @@ import (
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 
-	"github.com/stacklok/minder/internal/config"
+	"github.com/stacklok/minder/internal/constants"
 )
 
 // DatabaseConfig is the configuration for the database
@@ -69,36 +69,66 @@ func (c *DatabaseConfig) GetDBConnection(ctx context.Context) (*sql.DB, string, 
 
 // RegisterDatabaseFlags registers the flags for the database configuration
 func RegisterDatabaseFlags(v *viper.Viper, flags *pflag.FlagSet) error {
-	err := config.BindConfigFlagWithShort(
+	err := BindConfigFlagWithShort(
 		v, flags, "database.dbhost", "db-host", "H", "localhost", "Database host", flags.StringP)
 	if err != nil {
 		return err
 	}
 
-	err = config.BindConfigFlag(
+	err = BindConfigFlag(
 		v, flags, "database.dbport", "db-port", 5432, "Database port", flags.Int)
 	if err != nil {
 		return err
 	}
 
-	err = config.BindConfigFlagWithShort(
+	err = BindConfigFlagWithShort(
 		v, flags, "database.dbuser", "db-user", "u", "postgres", "Database user", flags.StringP)
 	if err != nil {
 		return err
 	}
 
-	err = config.BindConfigFlagWithShort(
+	err = BindConfigFlagWithShort(
 		v, flags, "database.dbpass", "db-pass", "P", "postgres", "Database password", flags.StringP)
 	if err != nil {
 		return err
 	}
 
-	err = config.BindConfigFlagWithShort(
+	err = BindConfigFlagWithShort(
 		v, flags, "database.dbname", "db-name", "d", "minder", "Database name", flags.StringP)
 	if err != nil {
 		return err
 	}
 
-	return config.BindConfigFlagWithShort(
+	return BindConfigFlagWithShort(
 		v, flags, "database.sslmode", "db-sslmode", "s", "disable", "Database sslmode", flags.StringP)
+}
+
+// GRPCClientConfig is the configuration for a service to connect to minder gRPC server
+type GRPCClientConfig struct {
+	// Host is the host to connect to
+	Host string `mapstructure:"host" default:"api.stacklok.com"`
+
+	// Port is the port to connect to
+	Port int `mapstructure:"port" default:"443"`
+
+	// Insecure is whether to allow establishing insecure connections
+	Insecure bool `mapstructure:"insecure" default:"false"`
+}
+
+// RegisterGRPCClientConfigFlags registers the flags for the gRPC client
+func RegisterGRPCClientConfigFlags(v *viper.Viper, flags *pflag.FlagSet) error {
+	err := BindConfigFlag(v, flags, "grpc_server.host", "grpc-host", constants.MinderGRPCHost,
+		"Server host", flags.String)
+	if err != nil {
+		return err
+	}
+
+	err = BindConfigFlag(v, flags, "grpc_server.port", "grpc-port", 443,
+		"Server port", flags.Int)
+	if err != nil {
+		return err
+	}
+
+	return BindConfigFlag(v, flags, "grpc_server.insecure", "grpc-insecure", false,
+		"Allow establishing insecure connections", flags.Bool)
 }
