+++ internal/config/server/config.go
@@ -37,7 +37,7 @@ type Config struct {
 	LoggingConfig LoggingConfig         `mapstructure:"logging"`
 	Tracing       TracingConfig         `mapstructure:"tracing"`
 	Metrics       MetricsConfig         `mapstructure:"metrics"`
-	Database      DatabaseConfig        `mapstructure:"database"`
+	Database      config.DatabaseConfig `mapstructure:"database"`
 	Identity      IdentityConfigWrapper `mapstructure:"identity"`
 	Auth          AuthConfig            `mapstructure:"auth"`
 	WebhookConfig WebhookConfig         `mapstructure:"webhook-config"`
